+++ support-plan-20210706/ChangeLog.md
@@ -1,3 +1,31 @@
+2024-09-13 Version: 2.0.0
+- Update API CreateTaskOrder: add param CreateUserId.
+- Update API CreateTaskOrder: add param IsUrgent.
+- Update API CreateTaskOrder: add param Overview.
+- Update API CreateTaskOrder: add param ProductCode.
+- Update API CreateTaskOrder: add param UrgentDescription.
+- Update API CreateTaskOrder: delete param CustomerRealName.
+- Update API CreateTaskOrder: delete param CustomerUserId.
+- Update API CreateTaskOrder: delete param ImportantDescription.
+- Update API CreateTaskOrder: delete param IsImportant.
+- Update API CreateTaskOrder: delete param ProductType.
+- Update API CreateTaskOrder: delete param ProductTypeName.
+- Update API CreateTaskOrder: delete param TaskTitle.
+- Update API CreateTaskOrder: update param OpenGroupId.
+- Update API ListDdTaskOrder: add param IsUrgent.
+- Update API ListDdTaskOrder: delete param CallerParentId.
+- Update API ListDdTaskOrder: delete param CallerType.
+- Update API ListDdTaskOrder: delete param CallerUid.
+- Update API ListDdTaskOrder: delete param OrderId.
+- Update API ListDdTaskOrder: delete param RequestId.
+- Update API ListDdTaskOrder: update param OpenGroupId.
+- Update API ListDdTaskOrder: update param PageNo.
+- Update API ListDdTaskOrder: update param PageSize.
+- Update API ListEnterpriseDingtalkGroupCustomerMembers: update param OpenGroupId.
+- Update API ListEnterpriseDingtalkGroupCustomerMembers: update param OpenGroupId.
+- Update API ListEnterpriseDingtalkGroupCustomerMembers: update response param.
+
+
 2024-09-05 Version: 1.0.4
 - Generated python 2021-07-06 for support-plan.